Move ColorSuggestionsBox to `@/components/colors`
