+++ src/components/colors/index.ts
@@ -1,3 +1,4 @@
-export * from './ColorGrid';
 export * from './ColorBlock';
+export * from './ColorGrid';
 export * from './ColorInput';
+export * from './ColorSuggestionsBox';
